Expand detection of MSVC
COMPILER_FAMILY_IS_MSVC was set according to detected build tools, but may have missed some (such as Ninja, which is available for Windows). CMake sets MSVC=1 when an MSVC compatible compiler is detected. It is possible that some versions of CMake might not set MSVC=1 in all cases when it should, but it is hard to imagine a false positive, so using two checks for COMPILER_FAMILY_IS_MSVC seems fair.